+++ RELEASE NOTES.md
@@ -1,3 +1,9 @@
+Version 1.1
+
+- Added object subscripting access methods to LSImageMap
+- Added fast enumeration support to LSImageMap
+- Images in LSImageMap are now sorted alphabetically
+
 Version 1.0.1
 
 - Fixed crash when loading TexturePacker v2 sprite sheets